# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Playwright & Selenium Learning Platform API
  description: |
    The Playwright & Selenium Learning Platform API allows developers to integrate learning features,
    access user data, manage webhooks, and build custom integrations.

    ## Authentication

    The API supports two authentication methods:

    ### API Keys
    Include your API key in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```

    ### OAuth 2.0
    Use OAuth 2.0 for user-authorized access:
    ```
    Authorization: Bearer YOUR_ACCESS_TOKEN
    ```

    ## Rate Limiting

    API requests are rate-limited based on your API key configuration (default: 1000 requests/hour).
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Total requests allowed per hour
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Time when the limit resets

    ## Scopes

    API keys and OAuth tokens require specific scopes for different operations:
    - `users:read` - Read user information
    - `users:write` - Update user information
    - `lessons:read` - Access lesson content
    - `lessons:write` - Create/update lessons
    - `quizzes:read` - Access quizzes
    - `quizzes:write` - Create/update quizzes
    - `webhooks:read` - View webhooks
    - `webhooks:write` - Manage webhooks
    - `api_keys:read` - View API keys
    - `api_keys:write` - Manage API keys

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /api-keys:
    get:
      tags:
        - API Keys
      summary: List API keys
      description: Get all API keys for the authenticated user
      operationId: listApiKeys
      security:
        - ApiKeyAuth: []
        - OAuth2: [api_keys:read]
      responses:
        '200':
          description: API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer

    post:
      tags:
        - API Keys
      summary: Create API key
      description: Create a new API key
      operationId: createApiKey
      security:
        - ApiKeyAuth: []
        - OAuth2: [api_keys:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      apiKey:
                        $ref: '#/components/schemas/ApiKey'
                      key:
                        type: string
                        description: The actual API key (shown only once)
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
  /api-keys/{keyId}:
    get:
      tags:
        - API Keys
      summary: Get API key
      description: Get a specific API key by ID
      operationId: getApiKey
      security:
        - ApiKeyAuth: []
        - OAuth2: [api_keys:read]
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          description: API key retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ApiKey'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - API Keys
      summary: Update API key
      description: Update an existing API key
      operationId: updateApiKey
      security:
        - ApiKeyAuth: []
        - OAuth2: [api_keys:write]
      parameters:
        - $ref: '#/components/parameters/KeyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApiKeyRequest'
      responses:
        '200':
          description: API key updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ApiKey'

    delete:
      tags:
        - API Keys
      summary: Revoke API key
      description: Revoke an API key
      operationId: revokeApiKey
      security:
        - ApiKeyAuth: []
        - OAuth2: [api_keys:write]
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
  /api-keys/{keyId}/rotate:
    post:
      tags:
        - API Keys
      summary: Rotate API key
      description: Generate a new key and revoke the old one
      operationId: rotateApiKey
      security:
        - ApiKeyAuth: []
        - OAuth2: [api_keys:write]
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          description: API key rotated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      apiKey:
                        $ref: '#/components/schemas/ApiKey'
                      key:
                        type: string
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Use your API key as a bearer token

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://api.playwright-learning.com/oauth/authorize
          tokenUrl: https://api.playwright-learning.com/v1/oauth/token
          scopes:
            users:read: Read user information
            users:write: Update user information
            lessons:read: Access lesson content
            lessons:write: Create/update lessons
            quizzes:read: Access quizzes
            quizzes:write: Create/update quizzes
            webhooks:read: View webhooks
            webhooks:write: Manage webhooks
            api_keys:read: View API keys
            api_keys:write: Manage API keys
  parameters:
    KeyId:
      name: keyId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        keyPrefix:
          type: string
        environment:
          type: string
          enum: [development, production]
        scopes:
          type: array
          items:
            type: string
        rateLimit:
          type: integer
        usageCount:
          type: integer
        lastUsedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CreateApiKeyRequest:
      type: object
      required:
        - name
        - environment
      properties:
        name:
          type: string
        environment:
          type: string
          enum: [development, production]
        scopes:
          type: array
          items:
            type: string
        rateLimit:
          type: integer
          default: 1000
        expiresAt:
          type: string
          format: date-time
        ipWhitelist:
          type: array
          items:
            type: string
        metadata:
          type: object
    UpdateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        rateLimit:
          type: integer
        isActive:
          type: boolean
        ipWhitelist:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              meta:
                type: object
                properties:
                  message:
                    type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...